+++ inform/__init__.py
@@ -1,5 +1,5 @@
-__version__ = '1.20.1'
-__released__ = '2020-01-18'
+__version__ = '1.20.2'
+__released__ = '2020-02-05'
 
 from .inform import (
     # inform utility functions and classes